+++ character/views/mixins.py
@@ -19,12 +19,11 @@ class CharacterContextMixin(ContextMixin, View):
         super().setup(request, *args, **kwargs)
         character_id = self.kwargs["character_id"]
         try:
-            self.character = cache.get(f"character{character_id}")
-            if not self.character:
-                self.character = Character.objects.get(id=character_id)
-                cache.set(f"character{character_id}", self.character)
-        except ObjectDoesNotExist:
-            raise Http404(f"Character [{character_id}] does not exist...")
+            self.character = cache.get_or_set(
+                f"character{character_id}", Character.objects.get(id=character_id)
+            )
+        except ObjectDoesNotExist as e:
+            raise Http404(f"Character [{character_id}] does not exist...") from e
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)