+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Versions follow [Semantic Versioning](https://semver.org/) (`<major>.<minor>.<pa
 * Use pylint as a linter
 
 ### Changed
+* Use `get_or_set()` where cache is used for better readability
 * Remove some errors raised by pylint
 
 ## v0.6.0 - 2024-02-03